Reuse base configuration with mocha-webpack (#7)
