+++ webpack.config.development.js
@@ -1,0 +1,3 @@
+const config = require('./webpack.config.base');
+
+module.exports = config;
